+++ inc/common.h
@@ -1,0 +1,14 @@
+#ifndef __COMMON_H
+#define __COMMON_H
+
+inline void __attribute__((always_inline))__enable_irq(void)
+{
+  __asm("CPSIE I");
+}
+
+inline void __attribute__((always_inline))__disable_irq(void)
+{
+  __asm("CPSID I");
+}
+
+#endif
