+++ inc/timer.h
@@ -26,10 +26,5 @@ void delayTimer(uint32_t mSec);
  * @param mSec delay time in approximate milli-seconds
  */
 void delayLoop(uint32_t mSec);
-/**
- * @brief The function Intializes Scheduler that uses Systick Timer as Timer source
- * 
- * @param useconds Frequency of the Systick Interrupt Generation in micro-seconds
- */
-void scheduler_Init(uint32_t useconds);
+
 #endif
